Bayer downscaling: fix downscaling for green pixels
For green pixels, we have to find the diamond which contains the targetted pixel.
Indeed, in Bayer matrix, the green pixels form diamonds:
RGR
GBG
RGR
To find which diamond contains the value to be interpolated, we compute the distance of the value to the centers of the two possible diamonds.
We then place ci, cj on the point which is just left of the top of the diamond, see the H here:
HGR
GBG
RGR