Contextual help: more user-friendly behaviour (cursor shape and message when no help)
We ignore the cursor shape changes that are not caused by us when we are in
contextual help mode, so that the cursor is not reset when we go towards the
left or the right of the screen.
We also display a message when the user does not click on a widget where we have help

Contextual help: open usermanual of same language as the one used in dt

If the usermanual is not available in this language, use english.

control: Use gboolean instead of C99 stdbool

Fixes #12380

Signed-off-by: Andreas Schneider <asn@cryptomilk.org>