colorpicker: fix colorpicker when used from proxy.
When used from a module, the colorpicker was not working if the
mode (POINT vs AREA) was changed by the module. In this case the
GUI was reset and the picker request was lost. We needed to click
again on the colorpicker to start drawing the area. This fixes
this long standing bad interaction by not reseting the GUI when
the picker mode was changed from a module (via the proxy).