Adaptive norm for nlmeans depending on pixel position (denoiseprofiled)
A pixel at a corner of an image will have a neighborhood less populated to compute
the norm.
Thus, as less pixels are used for the sum, we have to normalize by smaller numbers,
as the sum has to be normalized by the number of pixels involved.

Usually, P*P+1 pixels are involved, except on borders.