denoiseprofile: new option: whether to adapt transform to wb
After a first instance in color blending mode, the noise characteristics
of each RGB channel are melted together, so it makes no sense anymore to
use wb coeffs in the anscombe transform.
Yet, without such instance, it is much more effective in wavelets mode to
adapt the transform to wb coeffs.